moved the markup of invest log messages to the main process because it relies on a node package. #166.
